TIP-871: Move comment model to Akeneo\Enrichment
