+++ src/Akeneo/Pim/Enrichment/Component/Comment/Model/CommentInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\CommentBundle\Model;
+namespace Akeneo\Pim\Enrichment\Component\Comment\Model;
 
 use Doctrine\Common\Collections\ArrayCollection;
 use Symfony\Component\Security\Core\User\UserInterface;
