+++ src/Akeneo/Pim/Enrichment/Component/Product/Model/ProductInterface.php
@@ -10,7 +10,7 @@ use Akeneo\Tool\Component\StorageUtils\Model\ReferableInterface;
 use Akeneo\Tool\Component\Versioning\Model\VersionableInterface;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
-use Pim\Bundle\CommentBundle\Model\CommentSubjectInterface;
+use Akeneo\Pim\Enrichment\Component\Comment\Model\CommentSubjectInterface;
 use Pim\Component\Catalog\Model\EntityWithAssociationsInterface;
 use Pim\Component\Catalog\Model\EntityWithFamilyInterface;
 use Pim\Component\Catalog\Model\EntityWithFamilyVariantInterface;
