--- src/Pim/Bundle/CommentBundle/DependencyInjection/Compiler/ResolveDoctrineTargetModelPass.php
@@ -1,25 +1,0 @@
-<?php
-
-namespace Pim\Bundle\CommentBundle\DependencyInjection\Compiler;
-
-use Akeneo\Tool\Bundle\StorageUtilsBundle\DependencyInjection\Compiler\AbstractResolveDoctrineTargetModelPass;
-
-/**
- * Resolves doctrine ORM Target entities
- *
- * @author    Marie Bochu <marie.bochu@akeneo.com>
- * @copyright 2016 Akeneo SAS (http://www.akeneo.com)
- * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- */
-class ResolveDoctrineTargetModelPass extends AbstractResolveDoctrineTargetModelPass
-{
-    /**
-     * {@inheritdoc}
-     */
-    protected function getParametersMapping()
-    {
-        return [
-            'Pim\Bundle\CommentBundle\Model\CommentInterface' => 'pim_comment.entity.comment.class',
-        ];
-    }
-}
